Removed gst-common and inserted into gst-plugins
Revisited the plugin dependencies

In base, plugins reside in /usr/lib AND in /usr/lib/gstreamer-0.10
In base, the plugins in /usr/lib has suffix -0.10 and the rest none.

In gstreamer, the plugins residing in /usr/lib/ is REQUIRED, whereas the
plugins in /usr/lib/gstreamer-0.10 may be dynamically loaded and therefore
is OPTIONAL.

Removed the requirement for base plugins in the 2 fluendo plugins.

gst-plugins-good now only depends on gst-plugins-base *-0.10 plugins since
that is the plugins residing in /usr/lib.
If one includes plugins located in /usr/lib/gstreamer-0.10/ then it is up
to that person to verify that that plugin has its dependencies by manually
including the potential required plugins in /usr/lib/gstreamer-0.10