gstreamer: disable recipes for aarch64
The various gstreamer recipes don't build on aarch64. The 0.10.* series
has officially been unmaintained for almost 4 years [1], so getting
those to build would require a lot of work. The current 1.x recipes also
don't build, but a quick test shows that newer versions do.

In any case, the simplest for now is to blacklist an architecture where
the current gstreamer recipes have never been built for, to allow
meta/multimedia to get built by the buildbots.

[1] https://lists.freedesktop.org/archives/gstreamer-announce/2013-March/000273.html